refactor: LoginManager 위치 변경
- Interface, Sources에 있는 LoginManager 위치 변경
